fix: revert CopyToClipboard move, keep clipboard utility consolidation
The CopyToClipboard component consolidation revealed that the lib build
artifacts are significantly out of date (missing translation exports from
@apache-superset/core). This commit reverts the component move while
keeping the clipboard utility consolidation in @superset-ui/core.

The component move can be revisited once the lib build artifacts are
regenerated.

Changes:
- Revert src/components/CopyToClipboard to original implementation
- Restore types.ts file
- Remove CopyToClipboard from @superset-ui/core/components
- Keep clipboard utility (isSafari, copyTextToClipboard) in @superset-ui/core
- Keep test fixes for clipboard mock typing
