+++ superset-frontend/src/dashboard/components/menu/ShareMenuItems/ShareMenuItems.test.tsx
@@ -29,10 +29,9 @@ import { ComponentProps } from 'react';
 import { useShareMenuItems, ShareMenuItemProps } from '.';
 
 // Mock the clipboard utility
-const mockCopyTextToClipboard = jest.fn<
-  Promise<void>,
-  [() => Promise<string>]
->(() => Promise.resolve());
+const mockCopyTextToClipboard = jest.fn<Promise<void>, [() => Promise<string>]>(
+  () => Promise.resolve(),
+);
 jest.mock('src/utils/copy', () => ({
   __esModule: true,
   default: (getText: () => Promise<string>) => mockCopyTextToClipboard(getText),
